build(beta): per-component dirs for documenter + add beta documenter pass
Variant A: each flat BasicTable* part gets its own dir (basic-table-row, ...)
whose basename pascalCases to the component name, so the documenter names each
natively (no documenter-package change). docs.js gains a beta pass with glob
src/beta/*/*/index.tsx (verified via micromatch to match exactly the 7 component
dirs). Subpath entry re-exports the unchanged flat surface.
