+++ src/beta/basic-table-0.1/basic-table-header/index.tsx
@@ -1,0 +1,12 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+'use client';
+
+import { applyDisplayName } from '../internal/utils/apply-display-name';
+import { Header } from '../basic-table/internal';
+import { BasicTableProps } from '../basic-table/interfaces';
+
+applyDisplayName(Header, 'BasicTableHeader');
+
+export default Header;
+export type BasicTableHeaderProps = BasicTableProps.HeaderProps;
